first steps of new code generation is implemented. It supports Nullable and retryable methods. It can generate event and response parameters. It has also binds request , response and event together in templates so that relation can be seen and documented better

+++ hazelcast-code-generator/src/main/java/com/hazelcast/annotation/EventResponse.java
@@ -27,6 +27,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.METHOD)
 public @interface EventResponse {
-    short value();
+    int value();
 }
