+++ hazelcast-code-generator/src/main/java/com/hazelcast/annotation/Response.java
@@ -27,6 +27,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.SOURCE)
 @Target(ElementType.METHOD)
 public @interface Response {
-    short value();
+    int value();
 }
